completely refactor the logic as we got the core logic working, now to move everything onto the users courseProgressSchema and hold references to each course for each user.
TODO: Later once we add stripe move the course progress out of user
creation and put it where it belongs to go with the correct business
flow. Course details now retrieves the course details and the course
progress for that course the user is watching. later move this into
sessions and transactions to protect our data