get enrolled courses, we now populate more data for us to use in the frontend, in the toggleCompletion use new logic for toggling a lecture completed. add toggleCompletion method to controller. chnage completionPercentage virtual to handle new schema changes
Refactored the logic due to it becoming too messy and prone to errors,
the issue arose when i added more lectures to the course and then
because the courseProgress hadn't been updated with the new lectures
that broke the completionPercentage and other things. So now we handle
only completed lectures which we push to the completedLectures array for
said course or vice versa for un toggling. We also handle the case where
the course Progress dosen't exist(this will exisit when we purchase the
course). Moved into the course progress schema as it's own custom
function to keep controller tidier